+++ test/test_helper.bash
@@ -45,7 +45,7 @@ execute_with_unset_bats_vars() { # <command to execute...>
   "$@"
 }
 
-REENTRANT_RUN_PRESERVE+=(BATS_SAVED_PATH BATS_ROOT BATS_TEST_TAGS BATS_PARALLEL_BINARY_NAME)
+REENTRANT_RUN_PRESERVE+=(BATS_SAVED_PATH BATS_ROOT BATS_TEST_TAGS BATS_PARALLEL_BINARY_NAME BATS_LIBDIR)
 
 # call run with all BATS_* variables purged from the environment
 reentrant_run() { # <same args as run>