+++ docs/CHANGELOG.md
@@ -19,6 +19,8 @@ The format is based on [Keep a Changelog][kac] and this project adheres to
 
 ### Fixed
 
+* `install.sh` now works for deviating `lib/` dirs (like `lib32`,`lib64`) (#487)
+
 #### Documentation
 
 * clarify docker usage (#741)